Delete RegisteredUser
