+++ src/main/java/ebr/User.java
@@ -1,10 +1,14 @@
 package ebr;
 
 import java.io.Serializable;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
 
 public class User implements Serializable {
     protected String name;
     public String passHash, firstName, lastName, email;
+    private Set<Workout> workouts;
 
     protected User() {}
 
@@ -14,10 +18,26 @@ public class User implements Serializable {
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
+        this.workouts = new HashSet<>();
     }
 
     public String getName() {
         return name;
+    }
+
+    // Enrol (RegisteredUser) or Teach (Instructor) cert check is done in Workout.addUser
+    public void addWorkout(Workout w) {
+        if (w.addUser(this))
+            this.workouts.add(w);
+    }
+    public void removeWorkout(Workout w) {
+        if (workouts.remove(w)) {
+            w.removeUser(this);
+        }
+    }
+
+    public Set<Workout> getWorkouts() {
+        return Collections.unmodifiableSet(this.workouts);
     }
 
     @Override
